separate sb3 asset importer to it's own file
